refactor: remove v2 swapper naming (#3269)
* refactor: remove v2 swapper naming
